+++ Runtime/Core/Random/XoroShiroRandom.cs
@@ -69,8 +69,8 @@ namespace WallstopStudios.UnityHelpers.Core.Random
     )]
     [Serializable]
     [DataContract]
-    [ProtoContract]
-    [WProtoContract]
+    [ProtoContract(SkipConstructor = true)]
+    [WProtoContract(SkipConstructor = true)]
     public sealed partial class XoroShiroRandom
         : AbstractRandom,
             IEquatable<XoroShiroRandom>,